refactor: Move duplicate code into `getEntryPointFromMetafile()` function
