+++ packages/wrangler/src/deployment-bundle/bundle.ts
@@ -1,4 +1,3 @@
-import assert from "node:assert";
 import * as fs from "node:fs";
 import { builtinModules } from "node:module";
 import * as path from "node:path";
@@ -9,6 +8,7 @@ import tmp from "tmp-promise";
 import createModuleCollector from "../module-collection";
 import { getBasePath } from "../paths";
 import { dedent } from "../utils/dedent";
+import { getEntryPointFromMetafile } from "./entry-point-from-metafile";
 import { cloudflareInternalPlugin } from "./esbuild-plugins/cloudflare-internal";
 import { configProviderPlugin } from "./esbuild-plugins/config-provider";
 import { nodejsCompatPlugin } from "./esbuild-plugins/nodejs-compat";
@@ -394,23 +394,8 @@ export async function bundleWorker(
 		throw e;
 	}
 
-	const entryPointOutputs = Object.entries(result.metafile.outputs).filter(
-		([_path, output]) => output.entryPoint !== undefined
-	);
-	assert(
-		entryPointOutputs.length > 0,
-		`Cannot find entry-point "${entry.file}" in generated bundle.` +
-			listEntryPoints(entryPointOutputs)
-	);
-	assert(
-		entryPointOutputs.length < 2,
-		"More than one entry-point found for generated bundle." +
-			listEntryPoints(entryPointOutputs)
-	);
-
-	const { exports: entryPointExports, inputs: dependencies } =
-		entryPointOutputs[0][1];
-	const bundleType = entryPointExports.length > 0 ? "esm" : "commonjs";
+	const entryPoint = getEntryPointFromMetafile(entry.file, result.metafile);
+	const bundleType = entryPoint.exports.length > 0 ? "esm" : "commonjs";
 
 	const sourceMapPath = Object.keys(result.metafile.outputs).filter((_path) =>
 		_path.includes(".map")
@@ -418,7 +403,7 @@ export async function bundleWorker(
 
 	const resolvedEntryPointPath = path.resolve(
 		entry.directory,
-		entryPointOutputs[0][0]
+		entryPoint.relativePath
 	);
 
 	// A collision between additionalModules and moduleCollector.modules is incredibly unlikely because moduleCollector hashes the modules it collects.
@@ -440,7 +425,7 @@ export async function bundleWorker(
 
 	return {
 		modules,
-		dependencies,
+		dependencies: entryPoint.dependencies,
 		resolvedEntryPointPath,
 		bundleType,
 		stop: result.stop,
@@ -600,15 +585,6 @@ async function applyMiddlewareLoaderFacade(
 }
 
 /**
- * Generate a string that describes the entry-points that were identified by esbuild.
- */
-function listEntryPoints(
-	outputs: [string, ValueOf<esbuild.Metafile["outputs"]>][]
-): string {
-	return outputs.map(([_input, output]) => output.entryPoint).join("\n");
-}
-
-/**
  * Prefer modules towards the end of the array in the case of a collision by name.
  */
 export function dedupeModulesByName(modules: CfModule[]): CfModule[] {
@@ -619,9 +595,6 @@ export function dedupeModulesByName(modules: CfModule[]): CfModule[] {
 		}, {} as Record<string, CfModule>)
 	);
 }
-
-type ValueOf<T> = T[keyof T];
-
 /**
  * Process the given file path to ensure it will work on all OSes.
  *
